+++ src/main/java/com/robertx22/age_of_exile/aoe_data/database/perks/BigPerks.java
@@ -11,6 +11,7 @@ import com.robertx22.age_of_exile.database.data.stats.types.defense.MaxElemental
 import com.robertx22.age_of_exile.database.data.stats.types.generated.*;
 import com.robertx22.age_of_exile.database.data.stats.types.misc.DamageTakenToMana;
 import com.robertx22.age_of_exile.database.data.stats.types.offense.*;
+import com.robertx22.age_of_exile.database.data.stats.types.offense.crit.CriticalDamage;
 import com.robertx22.age_of_exile.database.data.stats.types.offense.crit.CriticalHit;
 import com.robertx22.age_of_exile.database.data.stats.types.offense.crit.SpellCriticalDamage;
 import com.robertx22.age_of_exile.database.data.stats.types.offense.crit.SpellCriticalHit;
@@ -41,6 +42,17 @@ public class BigPerks implements ISlashRegistryInit {
         PerkBuilder.bigStat("big_int", "Wisdom", new OptScaleExactStat(30, Intelligence.INSTANCE, ModType.FLAT));
         PerkBuilder.bigStat("big_dex", "Skill", new OptScaleExactStat(30, Dexterity.INSTANCE, ModType.FLAT));
         PerkBuilder.bigStat("big_str", "Power", new OptScaleExactStat(30, Strength.INSTANCE, ModType.FLAT));
+
+        PerkBuilder.bigStat("big_hp_res", "Hard Life",
+            new OptScaleExactStat(15, Health.getInstance(), ModType.LOCAL_INCREASE),
+            new OptScaleExactStat(1, new MaxElementalResist(Elements.Elemental), ModType.FLAT)
+        );
+
+        PerkBuilder.bigStat("big_eviscerator", "Eviscerator",
+            new OptScaleExactStat(3, AttackStyleDamage.MELEE, ModType.FLAT),
+            new OptScaleExactStat(10, CriticalHit.getInstance(), ModType.LOCAL_INCREASE),
+            new OptScaleExactStat(5, CriticalDamage.getInstance(), ModType.FLAT)
+        );
 
         PerkBuilder.bigStat("big_divine_light", "Divine Light",
             new OptScaleExactStat(5, SpellDamage.getInstance(), ModType.FLAT),
@@ -76,14 +88,44 @@ public class BigPerks implements ISlashRegistryInit {
             new OptScaleExactStat(200, DodgeRating.getInstance(), ModType.FLAT)
         );
 
+        PerkBuilder.bigStat("big_fury_bolt", "Furious Bolts",
+            new OptScaleExactStat(5, ProjectileDamage.getInstance(), ModType.FLAT),
+            new OptScaleExactStat(5, ProjectileSpeedStat.getInstance(), ModType.LOCAL_INCREASE),
+            new OptScaleExactStat(20, Strength.INSTANCE, ModType.FLAT)
+        );
+
+        PerkBuilder.bigStat("big_blood_thirst", "Blood thirst",
+            new OptScaleExactStat(10, Health.getInstance(), ModType.LOCAL_INCREASE),
+            new OptScaleExactStat(3, Lifesteal.getInstance(), ModType.FLAT),
+            new OptScaleExactStat(2, new AttackDamage(Elements.Physical), ModType.FLAT)
+        );
+
+        PerkBuilder.bigStat("big_bravery", "Soul of Bravery",
+            new OptScaleExactStat(20, Health.getInstance(), ModType.FLAT),
+            new OptScaleExactStat(10, Armor.getInstance(), ModType.LOCAL_INCREASE),
+            new OptScaleExactStat(10, DodgeRating.getInstance(), ModType.LOCAL_INCREASE)
+        );
+
         PerkBuilder.bigStat("big_frog", "Aspect of the Frog",
             new OptScaleExactStat(5, AttackStyleDamage.MELEE, ModType.FLAT),
             new OptScaleExactStat(10, DodgeRating.getInstance(), ModType.LOCAL_INCREASE),
             new OptScaleExactStat(10, Health.getInstance(), ModType.FLAT)
         );
 
+        PerkBuilder.bigStat("big_arena", "Master of the Arena",
+            new OptScaleExactStat(5, new AttackDamage(Elements.Physical), ModType.LOCAL_INCREASE),
+            new OptScaleExactStat(1, RegeneratePercentStat.HEALTH, ModType.FLAT),
+            new OptScaleExactStat(20, Strength.INSTANCE, ModType.FLAT)
+        );
+
+        PerkBuilder.bigStat("big_gladiator", "Master Gladiator",
+            new OptScaleExactStat(10, Accuracy.getInstance(), ModType.LOCAL_INCREASE),
+            new OptScaleExactStat(5, AttackSpeed.getInstance(), ModType.FLAT),
+            new OptScaleExactStat(20, Dexterity.INSTANCE, ModType.FLAT)
+        );
+
         PerkBuilder.bigStat("big_rat", "Aspect of the Rat",
-            new OptScaleExactStat(5, new AttackDamage(Elements.Physical), ModType.FLAT),
+            new OptScaleExactStat(5, new AttackDamage(Elements.Physical), ModType.LOCAL_INCREASE),
             new OptScaleExactStat(10, CriticalHit.getInstance(), ModType.LOCAL_INCREASE),
             new OptScaleExactStat(10, Dexterity.INSTANCE, ModType.FLAT),
             new OptScaleExactStat(10, Intelligence.INSTANCE, ModType.FLAT)
@@ -179,7 +221,7 @@ public class BigPerks implements ISlashRegistryInit {
         );
 
         PerkBuilder.bigStat("big_fighter", "Natural Fighter",
-            new OptScaleExactStat(3, new AttackDamage(Elements.Physical), ModType.FLAT),
+            new OptScaleExactStat(3, new AttackDamage(Elements.Physical), ModType.LOCAL_INCREASE),
             new OptScaleExactStat(3, AttackSpeed.getInstance(), ModType.FLAT),
             new OptScaleExactStat(20, Accuracy.getInstance(), ModType.FLAT),
             new OptScaleExactStat(20, Strength.INSTANCE, ModType.FLAT)
@@ -214,10 +256,21 @@ public class BigPerks implements ISlashRegistryInit {
             new OptScaleExactStat(1, RegeneratePercentStat.HEALTH, ModType.FLAT)
         );
 
-        PerkBuilder.bigStat("big_health_mana_cost", "Righteous Indignation",
+        PerkBuilder.bigStat("big_health_ms_mana_cost", "Righteous Indignation",
             new OptScaleExactStat(5, ReducedManaCost.getInstance(), ModType.FLAT),
             new OptScaleExactStat(10, Health.getInstance(), ModType.FLAT),
             new OptScaleExactStat(5, MagicShield.getInstance(), ModType.LOCAL_INCREASE)
+        );
+
+        PerkBuilder.bigStat("big_health_mana_cost", "Righteous Indignation",
+            new OptScaleExactStat(5, ReducedManaCost.getInstance(), ModType.FLAT),
+            new OptScaleExactStat(10, Health.getInstance(), ModType.LOCAL_INCREASE)
+        );
+        PerkBuilder.bigStat("big_faith_armor", "Armor of Faith",
+            new OptScaleExactStat(20, Armor.getInstance(), ModType.LOCAL_INCREASE),
+            new OptScaleExactStat(10, MagicShield.getInstance(), ModType.LOCAL_INCREASE),
+            new OptScaleExactStat(2, new ElementalResist(Elements.Elemental), ModType.FLAT)
+
         );
 
     }