+++ src/main/java/com/robertx22/age_of_exile/aoe_data/database/perks/ComboPerks.java
@@ -4,12 +4,14 @@ import com.robertx22.age_of_exile.database.OptScaleExactStat;
 import com.robertx22.age_of_exile.database.data.stats.types.defense.Armor;
 import com.robertx22.age_of_exile.database.data.stats.types.defense.DodgeRating;
 import com.robertx22.age_of_exile.database.data.stats.types.generated.AttackDamage;
+import com.robertx22.age_of_exile.database.data.stats.types.generated.ElementalResist;
 import com.robertx22.age_of_exile.database.data.stats.types.offense.AttackStyleDamage;
 import com.robertx22.age_of_exile.database.data.stats.types.offense.SpellDamage;
 import com.robertx22.age_of_exile.database.data.stats.types.resources.health.Health;
 import com.robertx22.age_of_exile.database.data.stats.types.resources.magic_shield.MagicShield;
 import com.robertx22.age_of_exile.database.data.stats.types.resources.mana.Mana;
 import com.robertx22.age_of_exile.database.data.stats.types.resources.mana.ManaRegen;
+import com.robertx22.age_of_exile.database.data.stats.types.spell_calc.ReducedManaCost;
 import com.robertx22.age_of_exile.database.registry.ISlashRegistryInit;
 import com.robertx22.age_of_exile.uncommon.enumclasses.Elements;
 import com.robertx22.age_of_exile.uncommon.enumclasses.ModType;
@@ -19,10 +21,21 @@ public class ComboPerks implements ISlashRegistryInit {
     @Override
     public void registerAll() {
 
+        PerkBuilder.stat("hp_mana_cost",
+            new OptScaleExactStat(2, ReducedManaCost.getInstance(), ModType.FLAT),
+            new OptScaleExactStat(4, Health.getInstance(), ModType.LOCAL_INCREASE)
+        );
+
+        PerkBuilder.stat("hp_res",
+            new OptScaleExactStat(5, Health.getInstance(), ModType.LOCAL_INCREASE),
+            new OptScaleExactStat(2, new ElementalResist(Elements.Elemental), ModType.FLAT)
+        );
+
         PerkBuilder.stat("spell_dmg_flat_mana",
             new OptScaleExactStat(5, SpellDamage.getInstance(), ModType.FLAT),
             new OptScaleExactStat(10, Mana.getInstance(), ModType.FLAT)
         );
+
         PerkBuilder.stat("spell_dmg_flat_mana",
             new OptScaleExactStat(3, SpellDamage.getInstance(), ModType.FLAT),
             new OptScaleExactStat(5, Mana.getInstance(), ModType.FLAT)