+++ src/main/java/com/robertx22/age_of_exile/database/data/stats/effects/resource/MagicStealEffect.java
@@ -4,7 +4,7 @@ import com.robertx22.age_of_exile.database.data.stats.Stat;
 import com.robertx22.age_of_exile.database.data.stats.effects.base.BaseDamageEffect;
 import com.robertx22.age_of_exile.saveclasses.unit.StatData;
 import com.robertx22.age_of_exile.uncommon.effectdatas.DamageEffect;
-import com.robertx22.age_of_exile.uncommon.effectdatas.EffectData;
+import com.robertx22.age_of_exile.uncommon.effectdatas.EffectUtils;
 
 public class MagicStealEffect extends BaseDamageEffect {
 
@@ -37,8 +37,7 @@ public class MagicStealEffect extends BaseDamageEffect {
 
     @Override
     public boolean canActivate(DamageEffect effect, StatData data, Stat stat) {
-        return effect.getEffectType()
-            .equals(EffectData.EffectTypes.BASIC_ATTACK);
+        return EffectUtils.isConsideredAWeaponAttack(effect);
     }
 
     private static class SingletonHolder {