+++ src/main/java/com/robertx22/age_of_exile/database/data/stats/effects/resource/ManaOnHitEffect.java
@@ -4,7 +4,7 @@ import com.robertx22.age_of_exile.database.data.stats.Stat;
 import com.robertx22.age_of_exile.database.data.stats.effects.base.BaseDamageEffect;
 import com.robertx22.age_of_exile.saveclasses.unit.StatData;
 import com.robertx22.age_of_exile.uncommon.effectdatas.DamageEffect;
-import com.robertx22.age_of_exile.uncommon.effectdatas.EffectData.EffectTypes;
+import com.robertx22.age_of_exile.uncommon.effectdatas.EffectUtils;
 
 public class ManaOnHitEffect extends BaseDamageEffect {
 
@@ -29,8 +29,7 @@ public class ManaOnHitEffect extends BaseDamageEffect {
 
     @Override
     public boolean canActivate(DamageEffect effect, StatData data, Stat stat) {
-        return effect.getEffectType()
-            .equals(EffectTypes.BASIC_ATTACK);
+        return EffectUtils.isConsideredAWeaponAttack(effect);
     }
 
 }